add profile tests
